+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Error.Reporting/Classes/Core/ConfigException.cs
@@ -1,0 +1,10 @@
+﻿namespace Krypton.Toolkit.Suite.Extended.Error.Reporting
+{
+    internal class ConfigException : Exception
+    {
+        public ConfigException(string message) : base(message)
+        {
+            
+        }
+    }
+}