+++ Documents/Help/Changelog.md
@@ -1,11 +1,14 @@
 =======
 
-## 2022-01-20 - Build 2201 - January 2022
+## 2022-02-03 - Build 2202 - February 2022
+* New `Krypton.Toolkit.Suite.Extended.Error.Reporting` module
+* Removal of unused 'using' statements
+* New `Krypton.Toolkit.Suite.Extended.Data.Visualisation` module (ScottPlot only)
 * Fix [#268](https://github.com/Krypton-Suite/Extended-Toolkit/issues/268), Crashes using Outlook Grid control
 * Implemented [#271](https://github.com/Krypton-Suite/Extended-Toolkit/issues/271), NuGet Package Information needs to have direct download links
 * Fixed issues where some canary NuGet packages were accidentally uploaded as stable packages
 * Refined build configurations
-* Updated standard toolkit to build `2201`
+* Updated standard toolkit to build `2202`
 
 =======
 
@@ -14,7 +17,7 @@
 * New `Krypton.Toolkit.Suite.Extended.InputBox` module
 * Fix toast 'Dismiss' button location
 * New `Krypton.Toolkit.Suite.Extended.Toast` module
-* Fixed [#245](https://github.com/Krypton-Suite/Extended-Toolkit/issues/245), 2108 Canary NuGet lists unecessary dependencies for packages
+* Fixed [#245](https://github.com/Krypton-Suite/Extended-Toolkit/issues/245), 2108 Canary NuGet lists unnecessary dependencies for packages
 * Fix possible `KryptonButton` text bugs
 * New icon
 * New `KryptonExceptionCaptureDialog` - it is now possible to display an exception in a window along with a typical messagebox